change authentification to support auth v2/v4
