+++ s3api/pkg/s3/liststorageclasses.go
@@ -20,15 +20,11 @@ import (
 	"github.com/opensds/multi-cloud/s3/pkg/model"
 	"github.com/opensds/multi-cloud/s3/proto"
 	"github.com/opensds/multi-cloud/s3api/pkg/common"
-	"github.com/opensds/multi-cloud/s3api/pkg/policy"
 	log "github.com/sirupsen/logrus"
 	"net/http"
 )
 
 func (s *APIService) GetStorageClasses(request *restful.Request, response *restful.Response) {
-	if !policy.Authorize(request, response, "storageclass:get") {
-		return
-	}
 	log.Info("Received request for storage classes.")
 
 	ctx := common.InitCtxWithAuthInfo(request)
