+++ s3api/pkg/s3/routbucketget.go
@@ -19,9 +19,6 @@ import (
 )
 
 func (s *APIService) RouteBucketGet(request *restful.Request, response *restful.Response) {
-	/*if !policy.Authorize(request, response, "bucket:get") {
-		return
-	}*/
 	if IsQuery(request, "acl") {
 		s.BucketAclGet(request, response)
 	} else if IsQuery(request, "uploads") {
